update style tag in +layout.svelte and search for Svelte SSR validation errors source
When the web page first loads, there are a few Svelte SSR validation errors that arise along the lines of "<x> is invalid inside <p>" (where x == nav, h1, div, etc). While I found the lines directly leading to the errors to appear, they did not coincide with the errors (ie the <nav> was definitively not in any <p>), so there is likely more tracing to do to find the root of the issue. For reference, the identified lines of code that directly led to the errors popping up are
 - src/routes/+page.svelte lines 54-60
 - src/routes/bar_graph/+page.svelte lines 63, 71-75
 - src/routes/bar_graph/+page.svelte lines 65-78